+++ playground/src/Components.fs
@@ -24,20 +24,37 @@ let Main () =
     let sortedFruits = Array.sort fruitArray // this creates a new array instance on each render
     let theme, setTheme = React.useState "light"
     Html.div [
-        prop.style [ 
-            if theme = "light" then 
-                style.backgroundColor "white"
-                style.color "black"
-            else 
-                style.backgroundColor "black"
-                style.color "white"
-        ]
-        prop.children [
-            Html.h1 "Fruit List"
-            Html.button [
-                prop.text "Add Grape"
-                prop.onClick (fun _ -> setFruitArray (Array.append fruitArray [| "Grape" |]))
+        Svg.svg [
+            svg.width 300
+            svg.height 120
+            svg.viewBox "0 0 300 120"
+            svg.children [
+                Svg.path [
+                    svg.d "M60,10 L60,110
+              M30,10 L300,10
+              M30,65 L300,65
+              M30,110 L300,110"
+                    svg.stroke "black"
+                    svg.strokeWidth 2
+                ]
+                Svg.text [
+                    svg.x 60
+                    svg.y 10
+                    svg.alignmentBaseline.hanging
+                    svg.text "A hanging"
+                ]
+                Svg.text [
+                    svg.x 60
+                    svg.y 65
+                    svg.alignmentBaseline.middle
+                    svg.text "A middle"
+                ]
+                Svg.text [
+                    svg.x 60
+                    svg.y 110
+                    svg.alignmentBaseline.baseline
+                    svg.text "A baseline"
+                ]
             ]
-            ToggleThemeButton (theme, setTheme)
         ]
     ]