+++ src/Feliz/CHANGELOG.md
@@ -16,7 +16,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
-## 3.1.1 - 2025-12-15
+## 3.1.1 - 2026-01-01
+
+### ✨ Added
+
+- Support for `svg.viewBox` property overload accepting raw string (by @Freymaurer)
 
 ### 🐛 Fixed